Project restructure (#35)
* Massive overhaul of project structure

The app.py file was growing bloated because it
had all of the routes, forms, etc. defined in
one place.

In lieu of this, I modularized it into the app/
directory in separate files. The new `app.py`
file is `gch.py`, which serves as the main
entrypoint for the app.

I haven't modified the gunicorn command yet
that runs the app, so production runs prob.
don't work yet.

* make gunicorn runs work for run.sh

* modify Dockerfile for new project structure / directory hierarchy